Fully provision a DocumentCloud vagrant box.
Starting with a base Ubuntu environment, install all DocumentCloud
dependencies, and configure them correctly.  At the end setup a
demo user and start up the various solr and cloud_crowd services.

Expanded the existing scripts to cover a complete installation.
Most scripts are still written in shell, with a few in Ruby.  Tied
them together with a Script class which takes care of error
handling and running them in the correct order for proper
operations.

Hopefully this method will prove to be fairly resilient and keep
us from having to implement a full-blown Chef/Puppet provisioning.