Fully provision a DocumentCloud vagrant box.
Starting with a base Ubuntu environment, install all DocumentCloud dependencies,
and configures them accordingly.  At the end install a demo user and
start up the various solr and cloud_crowd services.

Created a script runner and expanded the existing scripts, running them in
the correct order for proper operations.  Most scripts are still
written in shell, with a few in Ruby.

This method will *hopefully* prove to be fairly resilient and (arguably)
simpler than a full-blown Chef/Puppet provisioning.